requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.5
regex==2020.7.14
spark==0.2.1
textblob==0.15.3
tqdm==4.48.2
Werkzeug==1.0.1
